moved ThreadLocalByteArrayCache to its own class
